#98-UserIdTemp: Auto stash before merge of "feature/#98-UserIdTemp" and "origin/development"

+++ app/src/main/java/com/d_vide/D_VIDE/app/presentation/Followings/components/FollowingItem.kt
@@ -127,7 +127,7 @@ fun FollowDeleteButton(
             }
             if (isFollowing && isClicked){
                 "팔로잉하는 중".log()
-                viewModel.postFollow(userId.toInt())
+                viewModel.postFollow(userId)
             }
             if (!isFollowing){
                 viewModel.deleteFollow(followId.toInt())
